+++ CHANGELOG.md
@@ -48,3 +48,4 @@
 - [x] add apisix vars into ctx 2024-04-15
 - [x] plugin cors [cors](https://apisix.apache.org/zh/docs/apisix/plugins/cors/) 2024-04-16
 - [x] plugin request-validateion [request-validateion](https://apisix.apache.org/zh/docs/apisix/plugins/request-validateion/) 2024-04-18
+- [x] plugin fault-injection [fault-injection](https://apisix.apache.org/zh/docs/apisix/plugins/fault-injection/) 2024-04-19